This package is now pure ESM. Flat config is now required. ESLint 9.20.0 or later is required. Update dependencies
