+++ lib/index.d.ts
@@ -1,29 +1,17 @@
-import type { Linter } from  "eslint";
+import type { Linter } from "eslint";
 
 declare const _default: {
   configs: {
-    flat: {
-      recommended: Linter.Config;
-      common: Linter.Config;
-      import: Linter.Config;
-      sonarjs: Linter.Config;
-      prettier: Linter.Config;
-      react: Linter.Config;
-      reactHooks: Linter.Config;
-      reactRefresh: Linter.Config;
-      jsxA11y: Linter.Config;
-      testingLibrary: Linter.Config;
-      promise: Linter.Config;
-      unicorn: Linter.Config;
-      perfectionist: Linter.Config;
-      typescript: Linter.Config;
-    };
     recommended: Linter.Config;
     common: Linter.Config;
     import: Linter.Config;
     sonarjs: Linter.Config;
     prettier: Linter.Config;
     react: Linter.Config;
+    reactHooks: Linter.Config;
+    reactRefresh: Linter.Config;
+    jsxA11y: Linter.Config;
+    testingLibrary: Linter.Config;
     promise: Linter.Config;
     unicorn: Linter.Config;
     perfectionist: Linter.Config;
