+++ lib/rules/prettier.js
@@ -1,4 +1,4 @@
 /** eslint-plugin-prettier */
-module.exports = {
-    "prettier/prettier": "error",
+export default {
+  "prettier/prettier": "error",
 }
