+++ lib/rules/react-refresh.js
@@ -1,5 +1,5 @@
 /** eslint-plugin-react-refresh */
-module.exports = {
-    // https://github.com/ArnaudBarre/eslint-plugin-react-refresh
-    "react-refresh/only-export-components": "warn",
+export default {
+  // https://github.com/ArnaudBarre/eslint-plugin-react-refresh
+  "react-refresh/only-export-components": "warn",
 }
