+++ lib/rules/react.js
@@ -1,7 +1,7 @@
 /** eslint-plugin-react */
-module.exports = {
-    // https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/prop-types.md
-    "react/prop-types": "off",
-    // https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md
-    "react/react-in-jsx-scope": "off",
+export default {
+  // https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/prop-types.md
+  "react/prop-types": "off",
+  // https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md
+  "react/react-in-jsx-scope": "off",
 }
