+++ lib/rules/sonarjs.js
@@ -1,7 +1,7 @@
 /** eslint-plugin-sonarjs */
-module.exports = {
-    // https://sonarsource.github.io/rspec/#/rspec/S4144/javascript
-    "sonarjs/no-identical-functions": ["error", 5],
-    // https://sonarsource.github.io/rspec/#/rspec/S1135/javascript
-    "sonarjs/todo-tag": "warn",
+export default {
+  // https://sonarsource.github.io/rspec/#/rspec/S4144/javascript
+  "sonarjs/no-identical-functions": ["error", 5],
+  // https://sonarsource.github.io/rspec/#/rspec/S1135/javascript
+  "sonarjs/todo-tag": "warn",
 }
