+++ lib/rules/typescript.js
@@ -1,13 +1,13 @@
 /** @typescript-eslint-* */
-module.exports = {
-    // https://typescript-eslint.io/rules/no-use-before-define/
-    "@typescript-eslint/no-use-before-define": ["error"],
-    // https://typescript-eslint.io/rules/no-unused-vars/
-    "@typescript-eslint/no-unused-vars": [
-        "error"
-    ],
-    // https://typescript-eslint.io/rules/no-explicit-any/
-    "@typescript-eslint/no-explicit-any": "error",
-    // https://typescript-eslint.io/rules/no-empty-object-type
-    "@typescript-eslint/no-empty-object-type": "off"
+export default {
+  // https://typescript-eslint.io/rules/no-use-before-define/
+  "@typescript-eslint/no-use-before-define": ["error"],
+  // https://typescript-eslint.io/rules/no-unused-vars/
+  "@typescript-eslint/no-unused-vars": [
+    "error"
+  ],
+  // https://typescript-eslint.io/rules/no-explicit-any/
+  "@typescript-eslint/no-explicit-any": "error",
+  // https://typescript-eslint.io/rules/no-empty-object-type
+  "@typescript-eslint/no-empty-object-type": "off"
 }
