+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-nimbus-clean",
-  "version": "2.0.1",
+  "version": "3.0.0",
   "description": "A comprehensive linting solution that sweeps your code clean. Fly through your codebase with ease and precision!",
   "keywords": [
     "eslint",
@@ -11,6 +11,7 @@
     "react",
     "typescript"
   ],
+  "type": "module",
   "author": "dipiash",
   "license": "MIT",
   "main": "./dist/index.js",
@@ -48,66 +49,66 @@
     "prepare": "husky"
   },
   "devDependencies": {
-    "@commitlint/cli": "19.3.0",
-    "@commitlint/config-conventional": "19.2.2",
-    "@eslint/js": "^9.14.0",
-    "@release-it/bumper": "6.0.1",
-    "@release-it/conventional-changelog": "8.0.2",
-    "@typescript-eslint/eslint-plugin": "^8.14.0",
-    "commitizen": "4.3.0",
+    "@commitlint/cli": "19.8.0",
+    "@commitlint/config-conventional": "19.8.0",
+    "@eslint/js": "^9.22.0",
+    "@release-it/bumper": "7.0.2",
+    "@release-it/conventional-changelog": "10.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.27.0",
+    "commitizen": "4.3.1",
     "cz-conventional-changelog": "3.3.0",
-    "cz-git": "1.9.1",
-    "dotenv": "16.4.5",
-    "dotenv-cli": " 7.4.2",
-    "eslint": "^9.14.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-doc-generator": "^1.7.1",
-    "eslint-import-resolver-typescript": "^3.6.3",
-    "eslint-plugin-eslint-plugin": "^6.3.1",
+    "cz-git": "1.11.1",
+    "dotenv": "16.4.7",
+    "dotenv-cli": "8.0.0",
+    "eslint": "^9.22.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-doc-generator": "^2.1.2",
+    "eslint-import-resolver-typescript": "^4.2.2",
+    "eslint-plugin-eslint-plugin": "^6.4.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-n": "^17.13.1",
-    "eslint-plugin-perfectionist": "^4.1.2",
-    "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-promise": "^7.1.0",
-    "eslint-plugin-react": "^7.37.2",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.14",
-    "eslint-plugin-sonarjs": "^2.0.4",
-    "eslint-plugin-testing-library": "^6.4.0",
-    "eslint-plugin-unicorn": "^56.0.0",
-    "globals": "^15.12.0",
-    "husky": "9.1.6",
+    "eslint-plugin-n": "^17.16.2",
+    "eslint-plugin-perfectionist": "^4.10.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "eslint-plugin-sonarjs": "^3.0.2",
+    "eslint-plugin-testing-library": "^7.1.1",
+    "eslint-plugin-unicorn": "^57.0.0",
+    "globals": "^16.0.0",
+    "husky": "9.1.7",
     "markdown-toc": "1.2.0",
     "npm-run-all": "4.1.5",
-    "prettier": "^3.3.3",
-    "release-it": "17.2.1",
-    "syncpack": "12.3.2",
-    "typescript-eslint": "^8.14.0"
+    "prettier": "^3.5.3",
+    "release-it": "18.1.2",
+    "syncpack": "13.0.3",
+    "typescript-eslint": "^8.27.0"
   },
   "engines": {
     "node": "^14.17.0 || ^16.0.0 || >= 18.0.0"
   },
   "peerDependencies": {
-    "@eslint/js": "^9.14.0",
-    "typescript-eslint": "^8.14.0",
-    "@typescript-eslint/eslint-plugin": "^8.14.0",
-    "eslint": "^9.14.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "@eslint/js": "^9.22.0",
+    "typescript-eslint": "^8.27.0",
+    "@typescript-eslint/eslint-plugin": "^8.27.0",
+    "eslint": "^9.22.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-import-resolver-typescript": "^4.2.2",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-perfectionist": "^4.1.2",
-    "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-promise": "^7.1.0",
-    "eslint-plugin-react": "^7.37.2",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.14",
-    "eslint-plugin-sonarjs": "^2.0.4",
-    "eslint-plugin-testing-library": "^6.4.0",
-    "eslint-plugin-unicorn": "^56.0.0",
-    "globals": "^15.12.0",
-    "prettier": "^3.3.3"
+    "eslint-plugin-perfectionist": "^4.10.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "eslint-plugin-sonarjs": "^3.0.2",
+    "eslint-plugin-testing-library": "^7.1.1",
+    "eslint-plugin-unicorn": "^57.0.0",
+    "globals": "^16.0.0",
+    "prettier": "^3.5.3"
   },
   "peerDependenciesMeta": {
     "eslint": {
