+++ scripts/prepareTOCinReadme.sh
@@ -1,4 +1,1 @@
-markdown-toc -i README.md --maxdepth 10 &&
-markdown-toc -i docs/ja/README.md --maxdepth 10 &&
-markdown-toc -i docs/zh-tw/README.md --maxdepth 10 &&
-markdown-toc -i docs/hi/README.md --maxdepth 10
+markdown-toc -i README.md --maxdepth 10
